+++ src/main/java/de/maxhenkel/easyvillagers/blocks/IncubatorBlock.java
@@ -102,7 +102,7 @@ public class IncubatorBlock extends VillagerBlockBase implements EntityBlock, II
 
     @Override
     public RenderShape getRenderShape(BlockState state) {
-        return RenderShape.INVISIBLE;
+        return RenderShape.MODEL;
     }
 
     @OnlyIn(Dist.CLIENT)